rename(feature/#49): api 폴더명 apis로 변경
